CDAP-19817 feat: code refactored
