+++ app/cdap/components/FooterPanel/__test__/FooterPanel.test.tsx
@@ -52,16 +52,9 @@ describe('Test Footer Panel Component', () => {
   });
 
   it('Should render all the child elements correctly in the parent component', () => {
-    const parentElement = screen.getByTestId(/footer-panel-container/i);
     const wrapperElement = screen.getByTestId(/footer-panel-wrapper/i);
 
-    // Check if the parent element is rendered on screen
-    expect(parentElement).toBeInTheDocument();
-
-    // Check if the wrapper element is inside the parent element on the screen
-    expect(parentElement).toContainElement(wrapperElement);
-
-    const viewPanelTabParent = screen.getByTestId(/footer-panel-column-view-panel-tab-wrapper/i);
+    const viewPanelTabParent = screen.getByTestId(/footer-panel-column-view-panel-tab/i);
 
     // Check if the ViewPanelTabParent child component is rendered trough footer panel parent
     expect(wrapperElement).toContainElement(viewPanelTabParent);
